% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimisation.R
\name{gac}
\alias{gac}
\title{Generalised Additive Covariance Functions}
\usage{
gac(
  R,
  X,
  Emp_Cov,
  cov_func,
  param_eqns,
  param_init = NULL,
  loss = "WLS",
  smoothness_param = 0
)
}
\arguments{
\item{X}{Some container of covariates, e.g. a list with elements same size as \code{R}}

\item{cov_func}{A parametric covarianve function}

\item{param_eqns}{A list of equations for each parameter of \code{cov_func}}

\item{param_init}{Parameter values to initialise optimisaion}

\item{loss}{Chosen loss function. Options are:
\itemize{
 \item{WLS}{Weighted Leased Squares - weigthing by correlation}
 \item{WLSf}{Weighted Leased Squares - weigthing by all correlations}
 \item{LS}{Least Squares}
}}

\item{smoothness_param}{Parameter to control smoothness, is the weight that multiples
squared second derivative of smooth splines.}

\item{params}{R Separation matrix}

\item{Emp_CovEmpirical}{covaraiance to fit model to}
}
\value{
Returns an object of class \code{gac} for which methods will be written...
}
\description{
Function to fit a covairance function with generalised additive-type models
for parameters.
}
\details{
Fits models for generalised additive covariance functions. Work in progress!
}
\author{
Jethro Browell, \email{jethro.browell@glasgow.ac.uk}
}
\keyword{Covariance}
\keyword{Function}
