% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimisation.R
\name{gac_obj}
\alias{gac_obj}
\title{Objective functions for covariance functions}
\usage{
gac_obj(params, R, Emp_Cov, cov_func, loss = "WLS", ...)
}
\arguments{
\item{params}{A list of parameters to be passed to specified covairance function \code{cov_func}}

\item{R}{Separation matrix or vector}

\item{Emp_Cov}{Empirical covariance matrix against which covariance function is evaluated}

\item{cov_func}{A covariance function}

\item{loss}{Chosen loss function. Options are:
\itemize{
 \item{WLS}{Weighted Leased Squares - weigthing by cross-correlation}
 \item{WLSf}{Weighted Leased Squares - weigthing by all correlations}
 \item{LS}{Least Squares}
}}

\item{...}{Additional arguments passed to \code{cov_func}}
}
\description{
Function to evaluate a selected objective function for a given covariance
function and empirical covairance function.
}
\details{
Function that returns the value of the chosen loss function for given
empirical covariance matrix and covairance function for use in numerical methods.
}
\author{
Jethro Browell, \email{jethro.browell@glasgow.ac.uk}
}
\keyword{Covariance}
\keyword{Function}
