% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovarianceFunctions.R
\name{PowExp}
\alias{PowExp}
\title{Powered Exponential Covariance Function}
\usage{
PowExp(
  r = NULL,
  params = list(sigma = 1, theta = 1, gamma = 1),
  return_param_limits = F,
  optim_bound = F
)
}
\arguments{
\item{r}{Vector or matrix of separation distances}

\item{params}{A list of parameters with default \code{list(sigma=1,theta=1,gamma=1)}.
Parameters may be supplied as vectors of length equal to \code{length(r)}.}

\item{return_param_limits}{Boolean. If \code{TRUE} parameter limits are returned.}

\item{optim_bound}{Boolean. If \code{TRUE} parameters out side of limits do not result
in errors and instead a value of \code{Inf} is returned for these elements. For use in numerical
methods/optimisation.}
}
\value{
A vector of matrix the same size as \code{r} containing corresponding values of
the Powered Exponential Covariance Function, unless \code{return_param_limits=T} in which
case parameter limits are returned instead.
}
\description{
Functional form of the Powered Exponential Covariance Function
}
\details{
Function that returns the value of the Powered Exponential Covariance function.
}
\author{
Jethro Browell, \email{jethro.browell@glasgow.ac.uk}
}
\keyword{Covariance}
\keyword{Function}
