% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovarianceFunctions.R
\name{Seperable2}
\alias{Seperable2}
\title{Construction of a seperable covariance function}
\usage{
Seperable2(params, r1, r2, f1, f2, n_params1)
}
\arguments{
\item{params}{A vector of parameter values, a concationation of parameters for 
\code{f1} and \code{f2}.}

\item{r1}{Vector of matrix of separation distances for first separation variable}

\item{r2}{Vector of matrix of separation distances for second separation variable}

\item{f1}{Covariance function for first separation variable}

\item{f2}{Covariance function for second separation variable}

\item{n_params1}{Number of parameters to be passed to \code{f1}. Remainder are passed
to \code{f2}.}
}
\value{
A vector of matrix the same size as \code{r} containing corresponding values of
the Whittle-Matern covariance function.
}
\description{
A seperable covariance function, the product of two individual covariance functions.
}
\details{
Function that returns the value of the Whittle-Matern covariance function.
}
\author{
Jethro Browell, \email{jethro.browell@glasgow.ac.uk}
}
\keyword{Covariance}
\keyword{Function}
