% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovarianceFunctions.R
\name{Spherical}
\alias{Spherical}
\title{Spherical Covariance Function}
\usage{
Spherical(r, params = list(sigma = 1, theta = 1))
}
\arguments{
\item{r}{Vector or matrix of separation distances}

\item{params}{A list of parameters with default \code{list(sigma=1,theta=1)}.
Parameters may be supplied as vectors of length equal to \code{length(r)}.}
}
\value{
A vector of matrix the same size as \code{r} containing corresponding values of
the Spherical covariance function.
}
\description{
Functional form of the Spherical covariance function.
}
\details{
Function that returns the value of the Spherical covariance function.
}
\author{
Jethro Browell, \email{jethro.browell@glasgow.ac.uk}
}
\keyword{Covariance}
\keyword{Function}
